bufferpool: Detect use-after-free / concurrent operations (#1408)
This updates the buffer pool implementation to be mostly compatible with the existing YARPC interface, but adds use-after-free detection, as well as any concurrent misuse of the buffers.

In production, it will detect concurrent accesses to the same buffer
(Which are usually caused by a use-after-free because a buffer was
released too early). In tests, we can implement additional
use-after-free detection which will disable pooling, but will detect any
operations performed after the free.

This is inspired by the [TChannel protectmem pool](https://github.com/uber/tchannel-go/blob/master/frame_utils_test.go#L45) as well as the [M3 pool](https://godoc.org/github.com/m3db/m3x/pool).

Some differences:
 * TChannel uses the `mprotect` syscall, which requires using `unsafe` and page allocations.
 * m3x/pool is not API compatible with `*bytes.Buffer`
 * m3x/pool does not catch bad uses of the `[]byte` after they have been freed.

We'll likely move this out into a separate top-level repo that we can use between multiple projects, but let's start with an internal package to iron out the API and implementation.