Randomize the order of peers in peerlists.
Randomize the order of peers in peerlists.

Previously, peers would always be used in the order as input.

This can lead to non-uniform load on the peers in some cases.
For example:
- rpc callers are crash looping and repeatedly coming back up
  and hitting the first peer. (first peer will be hammered)
- many callers are started around the same time, and perform
  rpcs on a fixed schedule (each peer will be hammered in turn)

Instead, randomly shuffle peers as they are added.

By default, the random-number generator used for shuffling is
seeded with the current time in nanoseconds.

Support an Option for setting the random seed, to allow for
deterministic tests.

Support an Option for disabling the shuffle behavior, to allow
for greater clarity in some tests.