Add Ignored error code
This adds a `yarpcerrors.CodeIgnored` code to the errors API. This
is intended for applications/middleware to signal to the transport
that the inbound request should be ignored.

The TChannel transport will not respond to the request (black-hole)
causing clients to timeout. The client will never see the error
and therefore has no equivalent mapping to existing TChannel
error codes.

HTTP will return a `503: Service Unavailable` error to clients.