yarpc.Config: Option to disable observability middleware (#1421)
Summary: In order to allow some users to customize how they attach
observability middleware (e.g. users that want to log stats with shard
key).  This PR adds the ability to not add the built in observability
middleware on the dispatcher.  By default it's currently always added.
This will allow users to input their own middleware for observability
explicitly.  Since it would be a breaking change to remove it silently,
this needs to be an new option that explicitly disables the middleware..