[POC][No Land] Show how the self-assigned-sharder can be open-sourced
Summary: This is just a POC PR before I'm out.  This PR pulls the
self-assigned-sharder into yarpc as an /x/ package that can be reused.
Unfortunately when I pulled this out the tests couldn't come since they
were heavy integration tests that depend on a log of the Frontcar/Config
infra to set up properly.  But the core code is all here (and it works
internally).