Replace cover.sh with a Go script
Coverage files generated by `go test -coverprofile` take the form,

    ${importPath}/somefile.go:[coverage data goes here]

The cover.sh script attempts to filter out lines for files inside
ignored packages.

I noticed that this was taking really long on my computer. A couple
reasons for that:

- We build a *massive* regular expression to exclude lines
- The raw coverage data is massive in itself

Further, the regular expression isn't very accurate because it doesn't
delimit the packages. So it would ignore foo/bartest even if we're
trying to ignore foo/bar/baz.go.

This replaces the shell script with a Go program that attempts to be
both, more efficient and more accurate.

For reference, the number of lines in the raw coverage data before and
after this commit is,

```diff
 1608529 coverage.txt.raw.before
 1611009 coverage.txt.raw.after
```

The number of lines in the filtered coverage data before and after this
commit is,

```diff
 283713 coverage.txt.before
 326617 coverage.txt.after
```

Finally, I ran cover.sh twice before this commit, and cover.go twice
after this commit. The runtimes are,

```
Before  3:23    3:06
After   1:08    0:58
```

This decreases the build time by 2 minutes (on my machine).