make: Use Go 1.10's multi-package coverage (#1465)
This changes YARPC to rely on Go 1.10's multi-package coverage support
rather than the hacky cover.sh script we've had in the past.

The script still needs to exist to support the .nocover files but it's
now greatly simplified and we don't need to rely on gocovmerge,
parallel-exec. We've also not had to rely on the external cover tool
since Go 1.5.