Remove tracing from the HTTP outbound RoundTripper (#1470)
While refactoring the HTTP outbound to implement the `RoundTripper`
interface (#1444), tracing was forced on all calls made through
`RoundTrip(...)` and `Call(...)`. This creates an issue when using our 
internal tracing middleware, since we'd be reporting the same calls 
in two places.

Separating tracing support enables users to wrap the HTTP outbound 
with a custom tracing `RoundTripper`.