buildkite: Update dependencies automatically
This sets up automatic update of glide.lock for YARPC. It handles the
following scenarios:

- Nothing changed: Do nothing
- Only dependencies changed and tests continue to pass: Push directly to
  the target branch
- Dependencies and generated code changed: Create a pull request
- Tests failed: Create a pull request

For most cases, we will have an up-to-date glide.lock. When something
breaks, we'll get a PR and we'll have to figure out what went wrong and
fix it.

The following needs to be done once this change lands:

- Set up a yarpcbot account and give it write access to YARPC
- Generate SSH keys and a GitHub token for yarpcbot
- Upload the keys and token to the secret store
- Set up a scheduled build in BuildKite UI with the environment variable
  `YARPC_PIPELINE` set to `update-deps`