buildkite: Update dependencies automatically
This adds a script to etc/bin that updates the dependencies for this
repository automatically. It handles the following scenarios:

- Nothing changed: Do nothing
- Only dependencies changed and tests continue to pass: Push directly to
  the target branch
- Dependencies and generated code changed: Create a pull request
- Tests failed: Create a pull request

The plan is to set up a scheduled job in CI to run
pipeline-update-deps.yml daily.

For most cases, we will have an up-to-date glide.lock. When something
breaks, we'll get a PR and we'll have to figure out what went wrong and
fix it.

Once this change lands, we will have to set up a scheduled job in the
BuildKite UI with `YARPC_PIPELINE=update-deps `.