buildkite: Update docker-compose plugin
This updates the docker-compose plugin we use for BuildKite to the
latest version. There were no breaking changes between 1.5.2 and 1.8.4.

The change that caused a 2.0 to be released was because an internal key
was changed which would make it impossible to mix and match versions:
You couldn't use docker-compose#v1.8.4 with the version that had that
change. This required a major version bump. As long as all usages of the
plugin in the repo are using the same version, this is completely fine.