buildkite: Switch to dynamic pipeline
BuildKite supports the concept of Dynamic Pipelines. You can read more
about it [here][1] but the gist of it is: You can have a shell script
output the YAML for your pipeline instead of writing it by hand.

[1]: https://buildkite.com/docs/pipelines/defining-steps#dynamic-pipelines

This commit switches us over to a dynamic pipeline that lets us choose
between different pipeline.yml files based on an environment variable.

This change is motivated by the intent to introduce a scheduled job on
the YARPC pipeline that runs a different set of commands than the
default pipeline. Since individual jobs can have extra enviroment
variables specified, the scheduled job will specify a different
YARPC_PIPELINE than the default.

Once this change lands, we need to make the following change to the
pipeline upload step in the BuildKite UI.

```diff
-buildkite-agent pipeline upload
+.buildkite/pipeline.sh | buildkite-agent pipeline upload
```