Remove capacity assertions (#1512)
This walks back a minor part of a change I just landed. Since the
`Capacity` option is only setting the *initial* capacity of peer
lists, this method is fairly usesless. This is especially true as the
underlying structures freely expand as necessary.

This function was initially exposed for testing assertions outside of
the respective peer list packages. However, this will likely cause
confusion among users.

This preserves the behaviour exposed in #1505 and #1506.