http: Inbound.Stop should shutdown http.Server
The Inbound.Stop method currently stops the net.Listener, which stops
accepting new connections, but existing idle connections can
successfully make more requests.

Using Shutdown ensures we've closed all idle connections, and it waits
for all pending calls to complete.

The Shutdown method accepts a context, but we don't have a context
when stopping inbounds, so we have a couple of options:
 * Come up with an arbitrary timeout (e.g., 5 seconds)
 * Use `context.Background` and allow Stop to block forever.

Neither seems ideal, so I went with the arbitrary timeout for now.