Enable handlers to write response headers (#1533)
There is a defficiency in the current `ResponseWriter` API where
handlers and inbound middleware cannot write transport-level response
headers. Adding to the existing `ResponseWriter` interface would be a
breaking change so this introduces a new interface,
`ResponseMetaWriter`, that existing `ResponseWriter`s can optionally
be upcasted to.

HTTP, gRPC and TChannel `ResponseWriter`s will be updated in follow up
PRs.