Outbound: drop lifecycle, introduce chooser and dialer fields
This drops all of the lifecycle management which is now pushed to
`Dialer.Start(..)` and introduces the Dialer+URL and Chooser fields
which will used in a follow up.

This ignores the symmetric resquest/response which will be changed in
the following commit.

This also drop all v1 imports from the Outbound.