[spring/yarpcprotobuf] Add integration tests for generated streaming endpoints
Now that the gRPC transport layer has been migrated to the v2 APIs,
this adds integration tests for the generated streaming APIs in
the new `protoc-gen-yarpc-go` plugin.

This includes a test for each of the different gRPC handler types:
unary, client-side streaming, server-side streaming, and bidirectional
streaming.

Note that this integration test primarily concerns itself with the
`yarpcprotobuf` encoding layer. A more sophisticated integration test
will be added in a later change.