[spring] Introduce `yarpc.ClientProvider` (#1594)
This change introduces a `yarpc.ClientProvider` interface which is intended to
be used by generated clients. `yarpcclient.Provider` is a basic registry that
allows users to specify named clients and is expected to be initialized by `Fx`
for the general case.

This change includes the changes for protobuf generated clients to demonstrate
the interaction.