Randomize insertion order for fewest-pending-heap
This ensures that peers are randomly inserted into the heap among
equally scored peers, by randomly swapping 'last' with another peer on
insertion.

We've observed behaviour where batch deploys 'herd' towards a select
few instances. The theory is that when the FPH degenerates to
round-robin behaviour, batches of peers being inserted into the list
get added to the end of the ring. Relates to T2152727.