[spring] Add nil error checks to gRPC
Depending on encoding behaviour, it was previously possible for the
gRPC transport to panic since it always expected a non-nil
`yarpc.Buffer`, even on errors.

This ensures that we check the `yarpc.Request` and `yarpc.Buffer`
before sending information back over the wire.

Although this block of code was previously unidiomatic for go, this
preserves the behaviour of sending an error message and body. Error
handling will be refactored in a follow up.