Add yarpc.Dialer.Name to distinguish dialers to their transports
Given that we can use a shared `yarpc.Dialer` across our different
outbounds, i.e. all `http` outbounds can use the same `yarpc.Dialer`,
this proposes a new `Name` method on the `yarpc.Dialer` interface.

With this, we can distinguish between the different transports
and access the respective `yarpc.Dialer` in the construction of
our `peer.[Chooser|List]`.

An alternative to this would be to introduce something along the
lines of a `NamedDialer` type that embeds both a `Name` and
`Dialer` attributes. But seeing that we only have one
`yarpc.Dialer` per transport, I felt that we should at least
consider the following.