Add *Provider interfaces and Named types
This adds the `ChooserProvider`, `DialerProvider`, and `ListProvider`
interfaces referenced to in the previous commits.

And similar to the `yarpc.Dialer` change in the previous commit,
this adds two concrete types, `NamedList` and `NamedChooser` so
so that we can distinguish between the different peer list
implementations that are explicitly configured.

I wanted to demonstrate this as an approach as opposed to the `Name`
method on the interface(s) because there can be multiple implementations
of the same peer-list policy for a variety of different outbound
configurations. There might be two round-robin `[Chooser|List]`s for
outbounds `foo` and `bar`, for example, whereas there will only ever
be a single `yarpc.Dialer` for `http`.