Refactor yarpcthrift to match yarpc.EncodingProcedure (#1640)
Now that we have `yarpc.EncodingProcedure` and `yarpc.InboundCodec`, we separate the encoding-level handler and the encoding/decoding of requests/responses. This commit includes a `thriftCodec` that implements `yarpc.InboundCodec` and isolates the thrift handler to work with `wire.Value`s, instead of yarpc requests and responses.

From this, we have some issues from converting error information between a `yarpcerror.Status` and thrift exceptions, which will be revisited later, in a future commit.