Ensure protobuf and gRPC fully round-trips
This includes errors and regular successes.

Since grpc-go reserves its headers, application errors in yarpc must be
returned by the yarpcgrpc handler. We could have our protobuf encoding
handler return an error, but this breaks some of the abstractions
that we have established in yarpc (specifically that application errors
should be returned by the encoding handler as a `yarpc.Response` with
ErrorInfo set and return no error), and to ensure that we keep these
abstractions, we serialize an error at the protobuf encoding level and
de-serialize the error at the transport layer on the inbound side, and
vice-versa on the outbound side. While this may seem like unnecessary
work, establishing these abstractions allows other transports to work
with Protobuf.

grpc-go reserving their headers also means that it will be difficult for
yarpcgrpc to support Thrift and JSON. Currently, we plan to disallow
using gRPC with encodings other than Protobuf.