[spring] Errors (#1661)
* Introduce yarpcerror API

 - Remove Newf
 - Add ApplicationErrorInfo to `yarpc.Response`

* Add EncodeError to InboundErrorCodec (#1671)

* Added ErrorCodec to EncodingProcedure

* Implement EncodeError for yarpcjson

* Implement EncodeError for yarpcthrift

* Implement EncodeError for yarpcproto

* Refactor yarpcthrift to match the error api (#1667)

* Refactor yarpcgrpc to match new errors API (#1676)

* Refactor golang/protobuf -> gogo/protobuf

* Disallow json handlers from returning both response and error

* Refactor yarpchttp to match new API

* Remove WrapError, IsStatus in yarpcerror

* Ensure protobuf and gRPC fully round-trips

This includes errors and regular successes.

Since grpc-go reserves its headers, application errors in yarpc must be
returned by the yarpcgrpc handler. We could have our protobuf encoding
handler return an error, but this breaks some of the abstractions
that we have established in yarpc (specifically that application errors
should be returned by the encoding handler as a `yarpc.Response` with
ErrorInfo set and return no error), and to ensure that we keep these
abstractions, we serialize an error at the protobuf encoding level and
de-serialize the error at the transport layer on the inbound side, and
vice-versa on the outbound side. While this may seem like unnecessary
work, establishing these abstractions allows other transports to work
with Protobuf.

grpc-go reserving their headers also means that it will be difficult for
yarpcgrpc to support Thrift and JSON. Currently, we plan to disallow
using gRPC with encodings other than Protobuf.

* Add comment for protobuf handle

Since our interface for UnaryEncodingHandler takes interface{} and
returns interface{}, I added a comment to note what type it exactly
returns.

* Fix Thrift's error details behavior

When refactoring yarpcthrift to fit EncodingProcedure's API and the new
errors API, we moved Thrift exceptions to go into yarpc error's details.
While doing so, the exception would miss a step and wouldn't decode
correctly. This commit fixes this.

* Refactor yarpcthrift for new errors API

* Refactor yarpcobservability middleware to fit new errors API

A future change will leverage the new errors API to add well-defined
logs and stats.