glide: Add tools.go and move to imports
This moves tool dependencies to the `imports:` section of the
glide.yaml and adds a tools.go with blank imports of these dependencies.

This is needed to have all dependencies of these tools picked up
correcly. For example, BurntSushi/toml, which is now a dependency of
staticcheck, was not being picked up.

Note:
- kisielk/gotool was removed from the glide.yaml beacuse it wasn't being
  used directly.
- golang/protobuf was added because prometheus/client_model doesn't
  explicitly pin to it.