gen/proto: Use registered File Descriptors
The code generated by protoc-gen-yarpc-go includes the
FileDescriptorProto of the file and all its dependencies in the
generated code as a `var`. This is the Protobuf representation of the
`.proto` file itself.

As far as I can tell, we are attempting to replicate the behavior of
Protobuf with regards to the contents of this `var`. It's unclear why,
but this has been causing our `make generate` to fail for a month.

The version that protobuf generates includes the following additional
call,

    proto.RegisterFile($fileName)

This is a call to [proto.RegisterFile][1] (or the golang/protobuf
equivalent), which has a corresponding lookup function
[FileDescriptor][2].

  [1]: https://godoc.org/github.com/gogo/protobuf/proto#RegisterFile
  [2]: https://godoc.org/github.com/gogo/protobuf/proto#FileDescriptor

Rather than attempting to replicate the upstream behavior, this change
switches our generated code to rely on the lookup function to populate
our copy of the descriptor.