proto/internal: Delete golden test
This deletes the protoc-gen-yarpc-go internal golden test because it has
been in a semi-broken state for a while. The golden file and the actual
file diverged a long time ago, the golden file staying closer to what
the test generates, and recently even that broke.

The gist of the issue is that the test can't actually invoke proto
properly because it relies on looking up the FileDescriptor rather than
having the actual file.

This test is redundant anyway because we run `make generatenodiff` as
part of our tests.