dispatcher: Support changing app error log level
This uses the previously added machinery to support changing the level
at which inbound application errors are logged by the observability
middleware.

Note that even though the underlying system supports changing the log
level for success and failure cases, we're only adding support for
changing the application error log level for now.