Add failing test for pending heap
This adds a failing test for the fewest pending peer heap. Detailed
more in T2388437, a peer subscriber that was removed from the heap,
could cause a panic. During a simultaneous call to a peer and removal
of the same peer, the heap incorrectly assumed that the subscriber
still exists in the heap.