Verify pending heap subscriber exists in list
This change passes the failing test introduced in the previous
commit. When removing a peer from the pending heap list, we set the
index to negative one to signal it has been removed.

This prevents the case where a stale subscriber doesn't attempt to
notify the peer list of a peer that was already removed.

See T2388437 for more details.