Pending heap: Fix panic caused by stale subscribers (#1729)
During a concurrent peer call and removal of that peer from a list update, it
was possible for a subscriber to notify the fewest pending heap peer list of an
already removed peer. This caused an indexing panic when attempting to fix the
heap. This change ensures that we first validate that the subscriber is not
stale. See T2388437 for more details.

This PR adds a failing test and its corresponding fix in separate,
individually reviewable commits.