Release peer list lock after stale subscriber removal
In #1729, we introduced a change to remove panics some services were
seeing, specifically, when calling with a peer at the same time as a
removing it. Within that fix, we do not release the peer list lock
when attempting to remove a stale subscriber.

Similar to #1729, this is a small edge case, which is why it was left
uncovered for this time.

Related T2831133.