grpc: Repro for SendMsg holding on to response writer's bytes
YARPC uses gRPC differently to other systems,
 * YARPC handles marshalling, and passes in `[]byte`
 * It uses a "pass-through" codec since the data is already marshalled

This causes some surprising behaviour with `SendMsg`. When `SendMsg` is
normally used, a proto struct is passed in, gRPC marshals this into a
byte slice, and then holds on to the byte slice even after the `SendMsg`
function returns (it's added to a queue, that is eventually written to
the wire). Since we use a pass-through codec, the byte slice we pass in
to `SendMsg` is held on to even after the method returns.

This displays the issue by swapping out the `*bytes.Buffer` to a
`*bufferpool.Buffer`.

Today, this doesn't cause issues as the response writers don't seem to
be pooled.