thrift: Close req.Body after reading the request (#1740)
The Request.Body often has a Close method, which in TChannel's case,
will release memory associated with the inbound request. Currently,
the transport defers a close of the body when handling a request, but
this means that the inbound request body is not closed until after
the request has been fully handled.

This holds on to memory for much longer than required, which can reduce
the effectiveness of pooling in services which have many concurrent
requests.

Instead, change the Thrift encoding layer to close the request body
as soon as the request is read. For TChannel, this releases the frame
associated with the inbound request before we call any user handler.