make: Don't install lint dependencies by default
We currently always install lint dependencies even if `make lint` was
never called. This is a problem because newer versions of staticcheck
have a hard dependency on Go 1.12, which means we can't build and
install it for the Go 1.11 job, which doesn't perform linting.

This changes the makefiles to stop installing lint dependencies by
default. Instead, they will be installed as needed because the
corresponding lint jobs already declare a dependency on `$(ERRCHECK)`,
`$(GOLINT)`, and `$(STATICCHECK)`.