Add error details support over grpc/http (#1755)
This commit enables error details over grpc-protobuf and http-jsonpb. 

In the transport layer's error struct (`yarpcerrors.Status`), error details is backed by a `details` field of type `[]byte`. For gRPC, this is re-unmarshalled into a gRPC status and returned to the gRPC handler API. For HTTP, the details field is sent over the body *and* on a header with header key `grpc-error-details-bin` as is.

In the encoding layer's error struct (`pberror` in `encoding/protobuf` package), error details is backed by `[]proto.Message`. This error is marshaled into `[]byte` before it leaves the encoding layer on the inbound side, and a `yarpcerrors.Status`'s details field is unmarshaled to a `pberror` before it is returned to the user on the outbound side. This preserves any `proto.Message`s that are attached to the error as protobuf error details.