observability: emit error codes if the application error bit is set (#1739)
Currently, any error returned in a Protobuf handler will set the
application error bit. This prevents the metrics middleware from
emitting codes from `yarpcerrors`. Our metrics will always emit
`error:application_error` if the application error is set on the
`transport.ResponseWriter`.

This is fine in the Thrift case, since the application error bit is
only set for Thrift exceptions. However, the Protobuf encoding will
alway set the application error bit if an error is returned from a
handler. See:

 https://github.com/yarpc/yarpc-go/blob/bd70ffbd17e635243988ba62be97eebff738204d/encoding/protobuf/inbound.go#L70

This causes the metrics middleware to "hide" the richer error, always
defaulting to a caller error and generic "application_error" tag.

This change ensures that we only emit the generic tag if the
application error bit is set *and* we do not have a
`yarpcerrors.Status` as our error.

The alternative would be to change the Protobuf encoding to never set
the bit, since do not support the concept of application errors (ie
error details) in gRPC/Protobuf yet. However, this could impact
current HTTP/JSONpb users that are relient on the application error
header that HTTP sets.

Fixes T2942175. (More info in T2942175#59878297).