Add thrift procedures under second Fx value group
This adds thrift procedures under a second Fx value group. This could
for example, support an aggressive Thrift to Protobuf migration; we
may need to automatically generate Protobuf procedures based on the
output of the existing procedures generated with
`thriftrw-plugin-yarpc`.

This would enable us to depend on thrift generated procedures and
provide new ones in the `yarpcfx` value group.

I intentionally, skipped the CHANGELOG update since this is not
intended to be a user facing feature.