Privatize TChannel internal methods
The TChannel private peer type has some public methods.  These peers are provided to any peer list that retains or releases the peer for a given identifier, creating an API surface available to anyone who can cast the instance to an interface with the desired public method.  This change converts unnecessarily public methods to private to reduce the potential unintended API surface.