Pivot pendingheap from peerlist/v2 to abstractlist (#1829)
This change replaces the abstract list implementation for the pending request heap. The new abstract list tracks pending request counts itself instead of coordinating with the transport. #1828.

The most sensitive change to this implementation is that the heap no longer relies on peer.Status() at all (avoiding another potential deadlock), since the peer list sends the pending request count as an argument.

Since we’ve long since abandoned multi-dimensional peer scoring, this change also consolidates the score and pending request count into a single integer variable. At time of writing, we rely on the transport to provide relatively reliable connection statuses for each peer. Peer lists have a fail-fast mode and a wait-for-available mode, but don’t have a “here’s the best peer we’ve got, but you might not be able to connect” mode.