Introduce operation queue to abstract list
This change decouples the list updates from transport peer retain and release.

Once this change is made, the act of adding operations to the update queue is still synchronous with immediately flushing the updates.

All tests that depend on updates synchronously effecting retain and release will continue to pass, but the decoupling does prevent retain and release errors from being communicated back to the updater.

These retain and release errors are instead logged and this change deletes the obsolete test cases.