Reduce the public interface of peer lists
This change removes the publicly embedded concrete list implementation
from every peer list and instead exposes only the methods of peer.List
and peer.Chooser.

This is technically a breaking change.
There is potential for compatibility fallout with users that expected
additional methods.
However, we estimate that all these uses are bugs that should be fixed.