proto/grpc: Break dependency on examples
The tests for encoding/protobuf and transport/grpc have a dependency on
internal/examples/protobuf. This is undesirable because we want the
examples to be demonstrative, not test code.

Further, as we migrate to Go modules, we will likely want to move
examples into their own Go module, which we wouldn't tag, rather than
serving them as part of YARPC. That wouldn't be possible if we had test
code importing from that module.

This change copies the contents of internal/examples/protobuf to
internal/prototest/ and switches encoding/protobuf and transport/grpc to
use that in tests instead.