Add go.mod, go.sum files
This adds go.mod and go.sum files, generating them with `go mod init`
and `go mod tidy`, for the following paths:

- /: This is the Go module for YARPC itself.
- /internal/crossdock: This is the Go module for crossdock tests.
- /internal/examples: This is the Go module for examples. We should be
  free to move it to /examples in the future.

A nice effect of this separation is that YARPC itself does not declare
dependencies that are only needed for examples or crossdock, like
crossdock-go.

This deletes the glide.lock but leaves the glide.yaml in place to supply
constraints for legacy ecosystems.