generate.sh: Get gogo/protobuf location from Go modules
Since the source code for gogo/protobuf will no longer be present under
vendor/, we need to ask Go modules where it placed the source code so
that we can use it from protoc.

This calls `go mod download -json` to request that the code be
downloaded and gets its location with `jq`.

Note that since we'll already have called `go mod download` previously
to fetch all dependencies, this invocation of `go mod download` will
just use the cached copy.