Rename transport.Namer.Name() to TransportName()
oubound.Name() and outboundMiddleware.Name() both evoke the name of the
configured client, the callee service name.
This change renames these methods TransportName() to disambiguate their
meaning.
The fake transport name is "fake" and the outbound name does not
override this property.
If necessary, a future revision could add a transport name override
option to the fake transport, but for the time being, "fake" appears to
be sufficient for all existing tests.