yarpctest: Add ContextWithCall function
This adds a new ContextWithCall function which aims to ease testing of
functions which use yarpc.CallFromContext.

Previously, users would have to do something similar to the following,

    call, ctx := encoding.NewInboundCall(ctx)
    call.ReadFromRequest(&transport.Request{...})
    ...
    call.WriteToResponse(responseWriter)

With this change, the above becomes,

    ctx := yarpctest.ContextWithCall(ctx, &Call{...})
    ...

This relies on the prior change where we abstracted away the backing
metadata provider for Call into an inboundcall.Metadata interface.