encoding.Call: Refactor to override behavior
This refactors the implementation of encoding.Call (and consequently
yarpc.Call) to not rely on the internals of encoding.InboundCall.

Instead, a new internal interface is introduced that provides access to
the undelying metadata. We implement this interface around InboundCall
with a new unexported type.

By making the backing metadata an interface, we can provide an
alternative implementation from yarpctest.