observability: Log error messages from Protobuf handlers
This fixes an issue where all errors returned from a Protobuf handler
were logged uninformatively (identical to Thrift exceptions):

```
{ ... { ... "error": "application_error"}}
```

Unfortunately, all errors returned from Protobuf handlers are marked
as application errors and changing this behavior would be a breaking
change to HTTP/JSONpb users.

https://github.com/yarpc/yarpc-go/blob/bd70ffbd17e635243988ba62be97eebff738204d/encoding/protobuf/inbound.go#L71

With this PR we

1. Correctly log the error using `zap.Error(err)`, by detecting if we
   observed an error in addition to the application error bit being
   set.

2. Any `yarpcerror` created with details (including
  `encoding/protobuf.WithDetails`) are logged at the application error
  log level. This aligns with Thrift exceptions and would be the least
  surprising behavior for users (perhaps auto) migrating from Thrift
  exceptions to Protobuf error details.

Note that we made a similar workaround for metrics in #1739, where
regardless of the applciation bit being set or not, we'd emit the
correct metric.

Fixes T5590765