thrift: Return invalid argument error when `FromWire` fails (#1918)
This fixes a small edge case that came up with an internal tool that does not
guarantee sending valid requests :/

After an inbound request was unmarshaled, we would call `footype.FromWire(w)`.
If this call failed, we'd return the raw non-`yarpcerror` error. That behavior
reflects poorly on server metrics as it's deemed an unknown server error.

This change wraps the error in a `yarpcerror` using `CodeInvalidArgument`. This
aligns with the other errors returned from the Thrift and Protobuf encoding
layers.

Fixes #1909

For the majority of RPC use-cases, this should never be an issue. 
- [x] Description and context for reviewers: one partner, one stranger
- [x] Entry in CHANGELOG.md