abstractlist: Add default Choose timeout
This adds a default `Choose` timeout to the `abstractlist` so that
streaming calls with peer choosers do not require timeouts. If no
deadline is detected on the context, the peer list adds a timeout
(default of 500ms) instead of erroring out.

The added timeout is still required so that we do not wait
indefinitely for a peer to be available.