Fix gRPC streaming when used with the direct peer chooser (#1923)
We have an issue in the gRPC streaming outbound where we call OnFinish on the chosen peer, after the initial request was made.

The direct peer list uses places its "ReleasePeer" call in the "OnFinish" function, which the gRPC transport is prematurely calling.

Therefore, when streaming is used with the direct peer list this causes the flapping behavior, since the connection is being immediately closed. Other peer lists do not have this issue since peer lists retain the peers until they are actively removed from an updater.