Integrate checker for generated Protobuf Fx clients
This adds the restriction check to Protobuf clients. The important
change is in:

- `encoding/protobuf/protoc-gen-yarpc-go/internal/lib/lib.go`

Similar to failed client config lookups, we choose to fail loudly /
panic when we detect a restricted combination.

https://github.com/yarpc/yarpc-go/blob/4f4ed48bcde86e8709d248ddd0fd2712437874f3/dispatcher.go#L283

The other changes are as a result of `SUPPRESS_DOCKER=1 make generate`

Note that this only affects Fx clients.