observability: Override responses with context errors
THe majority of code here is identical to that introduced and
validated in D4334363.

Here, we add a context override function that determines whether or
not to throw away the user's response, in favour of the context
error. This ensures that YARPC metrics better reflect timeouts,
including higher correlation with sidecar/proxy metrics.