tchannel: Drop redundant "handler failed" log
This drops the "handler failed" log in the TChannel transport. This
log was unnecessarily added when we were increasing observability
around TChannel internal errors in #1561.

The context error override in #1930 ensures that makes this log
redundant as richer information exists in observability logs,
including latency and request attributes.

Furthermore, we've had issues with this log since the latency is
included in the message and makes aggregation extremely difficult.

Ref T5802517