introspection: Expose internal introspection APIs
Currently, it's not possible for users to piggyback off of YARPC's
existing debug pages if they implement a custom inbound, outbound or
peer chooser/list. This small change moves the respective types a top
level `api/x/introspection` package, for consumption.

Like all `x` packges in YARPC, these APIs are subject to change, as
documetned in the README.