Propagate application errors metadata for the Thrift encoding
Adds a new field on the transport.Response structure to hold the application error metadata. The metadata is composed of the error itself, the error name and the related YARPC code. The error name is the name of the Thrift exception extracted from the generated code. This commit only sets the metadata for the Thrift encoding.